pyproject.toml
[tool.poetry]
name = "montemirage"
version = "0.1.0"
description = ""
authors = ["Roman <romantovt31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^4.9.2"
requests = "^2.28.1"
discord-py = "^2.1.0"
python-dotenv = "^0.21.0"
peewee = "^3.15.4"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
flake8-annotations = "^2.9.1"
wemake-python-styleguide = "^0.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
